Add provider id to report data (#417)
* Model updates to track manifests and timestamps for bills
* Add provider_id column to aws cost entry bill

* Also allow providers to re-use billing and auth resources

* Call masu to delete provider report data

+++ koku/api/provider/provider_manager.py
@@ -18,8 +18,11 @@
 
 import logging
 
+import requests
+from django.conf import settings
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from django.db import transaction
+from requests.exceptions import ConnectionError
 
 from api.provider.models import Provider
 
@@ -40,8 +43,9 @@ class ProviderManager:
 
     def __init__(self, uuid):
         """Establish provider manager database objects."""
+        self._uuid = uuid
         try:
-            self.model = Provider.objects.all().filter(uuid=uuid).get()
+            self.model = Provider.objects.all().filter(uuid=self._uuid).get()
         except (ObjectDoesNotExist, ValidationError) as e:
             raise(ProviderManagerError(str(e)))
 
@@ -65,11 +69,40 @@ class ProviderManager:
             authentication_model = self.model.authentication
             billing_source = self.model.billing_source
 
-            authentication_model.delete()
-            if billing_source:
+            auth_count = Provider.objects.exclude(uuid=self._uuid)\
+                .filter(authentication=authentication_model).count()
+            billing_count = Provider.objects.exclude(uuid=self._uuid)\
+                .filter(billing_source=billing_source).count()
+            # Multiple providers can use the same role or bucket.
+            # This will only delete the associated records if it is the
+            # only provider using them.
+            if auth_count == 0:
+                authentication_model.delete()
+            if billing_source and billing_count == 0:
                 billing_source.delete()
+            try:
+                self._delete_report_data()
+            except ConnectionError as err:
+                LOG.error(err)
+                LOG.warning(('The masu service is unavailable. '
+                             'Unable to remove report data for provider.'))
             self.model.delete()
+
             LOG.info('Provider: {} removed by {}'.format(self.model.name, current_user.username))
         else:
             err_msg = 'User {} does not have permission to delete provider {}'.format(current_user, str(self.model))
             raise ProviderManagerError(err_msg)
+
+    def _delete_report_data(self):
+        """Call masu to delete report data for the provider."""
+        LOG.info('Calling masu to delete report data for provider %s',
+                 self.model.id)
+        params = {
+            'schema': self.model.customer.schema_name,
+            'provider': self.model.type,
+            'provider_id': self.model.id
+        }
+        # Delete the report data for this provider
+        delete_url = settings.MASU_BASE_URL + settings.MASU_API_REPORT_DATA
+        response = requests.delete(delete_url, params=params)
+        LOG.info('Response: %s', response.json())
